# backend/src/test/resources/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AiForDev API",
    "version": "0.1.0",
    "description": "REST API for the AiForDevelopers platform"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local backend"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/auth/oauth2/google/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Google OAuth2 callback",
        "operationId": "googleOAuthCallback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Authorization code from Google"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend with tokens",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "{FRONTEND_URL}/auth/callback?accessToken=...&refreshToken=...&email=..."
              }
            }
          }
        }
      }
    },
...
    "/api/calendar/google/callback": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Google Calendar OAuth2 callback",
        "operationId": "googleCalendarCallback",
        "security": [],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Authorization code from Google"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend with status"
          }
        }
      }
    },
    "/api/calendar/google/status": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Get Google Calendar connection status",
        "operationId": "getCalendarConnectionStatus",
        "responses": {
          "200": {
            "description": "Connection status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarConnectionStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid JWT"
          }
        }
      }
    },
    "/api/calendar/google/events": {
      "get": {
        "tags": [
          "calendar"
        ],
        "summary": "Get busy slots from Google Calendar",
        "operationId": "getCalendarBusySlots",
        "parameters": [
          {
            "name": "timeMin",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Start of the time period"
          },
          {
            "name": "timeMax",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "End of the time period"
          }
        ],
        "responses": {
          "200": {
            "description": "List of busy time slots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarBusySlotsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Calendar not connected or invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "notConnected": {
                    "summary": "Calendar not connected",
                    "value": {
                      "message": "Google Calendar is not connected"
                    }
                  },
                  "invalidParams": {
                    "summary": "Invalid query parameters",
                    "value": {
                      "message": "timeMin and timeMax are required"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid JWT"
          }
        }
      },
      "post": {
        "tags": [
          "calendar"
        ],
        "summary": "Create an event in Google Calendar",
        "operationId": "createCalendarEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCalendarEventRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarEventResponse"
                }
              }
            }
          },
          "400": {
            "description": "Calendar not connected or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "notConnected": {
                    "summary": "Calendar not connected",
                    "value": {
                      "message": "Google Calendar is not connected"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid JWT"
          }
        }
      }
    },
    "/api/calendar/google/events/{googleEventId}": {
      "delete": {
        "tags": [
          "calendar"
        ],
        "summary": "Delete an event from Google Calendar",
        "operationId": "deleteCalendarEvent",
        "parameters": [
          {
            "name": "googleEventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Google Calendar event ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Event deleted successfully"
          },
          "400": {
            "description": "Calendar not connected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "notConnected": {
                    "summary": "Calendar not connected",
                    "value": {
                      "message": "Google Calendar is not connected"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — missing or invalid JWT"
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "notFound": {
                    "summary": "Event not found",
                    "value": {
                      "message": "Event not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "CalendarConnectionStatusResponse": {
        "type": "object",
        "properties": {
          "connected": {
            "type": "boolean",
            "description": "Whether Google Calendar is connected"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Google Calendar email address"
          }
        },
        "required": [
          "connected"
        ]
      },
      "CalendarBusySlot": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time",
            "description": "Start of the busy slot"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "description": "End of the busy slot"
          }
        },
        "required": [
          "start",
          "end"
        ]
      },
      "CalendarBusySlotsResponse": {
        "type": "object",
        "properties": {
          "busy": {
            "type": "array",
            "description": "List of busy time slots",
            "items": {
              "$ref": "#/components/schemas/CalendarBusySlot"
            }
          }
        },
        "required": [
          "busy"
        ]
      },
      "CreateCalendarEventRequest": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "string",
            "description": "Event title/summary"
          },
          "description": {
            "type": "string",
            "description": "Event description"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "description": "Event start time"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "description": "Event end time"
          }
        },
        "required": [
          "summary",
          "start",
          "end"
        ]
      },
      "CalendarEventResponse": {
        "type": "object",
        "properties": {
          "googleEventId": {
            "type": "string",
            "description": "Google Calendar event ID"
          },
          "summary": {
            "type": "string",
            "description": "Event title/summary"
          },
          "description": {
            "type": "string",
            "description": "Event description"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "description": "Event start time"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "description": "Event end time"
          }
        },
        "required": [
          "googleEventId",
          "summary",
          "start",
          "end"
        ]
      }
    }
  }
}